requirements.txt
tvdatafeed @ git+https://github.com/rongardF/tvdatafeed.git
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
